Update configure machinery and add files to hold Dyninst analysis
The analysis uses tooling in Dyninst to do the analysis. However,
Dyninst is not available for some architectures. Needed to make the
configuration machinery only enable the Dyninst-based analysis for
supported machines such as x86-64 and aarch64.

The analysis.cxx is not currently connected to anything and doesn't do
any analysis.  For the moment it uses a couple Dyninst functions
verify that Dyninst libraries are being linked in.  This has been
tested to compile on both architectures that support Dyninst (x86_64)
and do not support Dyninst (32-bit Arm).