Do liveness analysis after location information set and use regno info
Initially the call to the liveness analysis was being done before
the location information was setup.  Relocated the call after the
location information has been setup.

Corrected liveness analysis code to use the register number.  Also
added print out the function the liveness analysis is being run on.
Note that for probes on inlined functions this name will be different
than the probe point function name, it will be the actual function
that the code was inlined into.