PR28923: dtrace: Use hash-based scheme for predictable file generation
commit c245153 ("dtrace: Allow for reproducible .o file builds.")
introduced a condition where 2 dtrace processes can race when
generating the same file. Since both processes now use the same
temporary file name, one may delete the temporary .c file the other
is still processing:

    --------------------------------------------------------------------
    user@host:~/foo$ make -j2
    dtrace -o foo.out -G -s /dev/null
    dtrace -o foo.out -G -s /dev/null
    Traceback (most recent call last):
      File "/usr/bin/dtrace", line 455, in <module>
        sys.exit(main())
      File "/usr/bin/dtrace", line 440, in main
        os.remove(fname)
    FileNotFoundError: [Errno 2] No such file or directory: 'foo.out.dtrace-temp.c'
    make: *** [Makefile:4: ../foo/foo.out] Error 1
    --------------------------------------------------------------------

This can happen when a Makefile processes a pattern rule for two different
targets that happen to map to the same file, but addressed by different
relative paths. I discovered this in a real world case involving libvirt,
but here's a contrived reproducer:

  --------------------------------------------------------------------
  all: foo.out ../$(basename $(CURDIR))/foo.out

  %.out:
        dtrace -o foo.out -G -s /dev/null

  clean:
        rm -f foo.out
  --------------------------------------------------------------------

It would be ideal if we could inject a null .file directive, then we could
just use a mkstemp() file and keep the build reproducible by avoiding a
record of the source file path in the binary at all, but I can't find a
straightforward way of passing a .file through to the assembler. So,
instead, let's create a reproducible filename by building a hash of the
input and output paths. Note: this still leaves open a race in the case of
2 dtrace processes with identical input/output paths. But, at least in my
testing, GNU Make is smart enough to detect this case and not create
duplicate jobs.

Fixes: Commit c245153 ("dtrace: Allow for reproducible .o file builds.")
Signed-off-by: dann frazier <dann.frazier@canonical.com>