sys/sdt.h: set x86-64 STAP_SDT_ASM_CONSTRAINT back to "nor"
It turns out the kernel and some other sdt consumers haven't learned
how to use %xmm registers in sdt operands.  So under this duress, stap
will go back to the old school integer register set "nor" as a
default.  We'll revisit this in the future, though this egg might not
turn into a chicken.

https://bugzilla.redhat.com/show_bug.cgi?id=2058908