PR28958: Fix tapset macros to allow nfsd-trace.stp and task_paths.stp to work
The task_dentry_path function in the tapset was getting read faults
instead of returning valid strings describing a filesystem path.  This
caused the nfsd-trace.stp example and task_path.stp in the testsuite
to not function properly.

The root cause of the problem was some cast operators in macros were
not accessing the proper struct debuginfo due to a missing "kernel" in
the cast operator.  The cast operators were corrected and scripts
using the task_dentry_path function now function properly.