transport: fix subbuf size to PAGE_SIZE
High-message-rate stap scripts more easily lose message synch or bog
down if the subbuf size is large.  PAGE_SIZE appears to be a sweet
spot, so let's fix that.  (At least one subbuf is used per probe hit
that produces output.  Allocation occurs at the subbuf granularity, so
making it smaller is apparently of no advantage.)  stap -s and
free-memory still affect transport memory allocation, but only as to
the number of subbufs.