Testsuite:  Prevent the hwcaps based dynamic loader search
The s390x syscall testsuite started to experience a problem where
the test logs were flooded with hundreds of newfstatat and openat
syscalls.  The reason for this was that the dynamic loader was
searching for shared objects based on hwcaps:

https://sourceware.org/pipermail/libc-alpha/2020-May/113757.html

This doesn't happen if the LD_LIBRARY_PATH isn't set.  Another
approach to avoid the flood would be to export LD_HWCAP_MASK=0.
This patch only  unsets LD_LIBRARY_PATH because it looks good
enough.

Instead of doing full autoreconf, which, in my case, would
produce a huge messy patch, this only is a targetted change
to the testsuite Makefile.{am,in}.

This update doesn't seem to break the dyninst part of the
testsuite, regardless of systemtap configure --prefix.