Extract the exit_reason from trace_kvm_exit vcpu argument on newer kernels
For x86_64 processors newer kernels change where the exit_reason
information is located.  In older kernels the exit_reason was a
parameter for the trace_kvm_exit.  For the newer kernels exit_reason
is a field buried in a member field of vcpu argument.  Making
kvm_service_time.stp pick the appropriate location for exit_reason.