PR29507: generalize sample python tapset for loose python{2,3} library versions
We can rely on stap 4.2+'s probe-context passing to functions to make
it unnecessary to decorate each @cast() with a libpython path name.
This lets these tests work on a range of python libraries.

These helper functions really should go into the standard python tapset,
rather than sit here in the examples, but that's for later.